+++ server.js
@@ -1,40 +1,8 @@
 const express = require('express');
 const cors = require('cors');
 const morgan = require('morgan');
-const { pick } = require('lodash');
+const { initCommuneFields, initCommuneFormat, formatCommune } = require('./lib/communeHelpers');
 const db = require('./lib/db');
-
-function initCommuneFields(req, res, next) {
-  if (req.query.fields) {
-    req.fields = new Set(req.query.fields.split(','));
-  } else {
-    req.fields = new Set(['nom', 'codeInsee', 'codesPostaux', 'centre', 'surface']);
-  }
-  req.fields.add('codeInsee', 'nom');
-  next();
-}
-
-function initCommuneFormat(req, res, next) {
-  req.outputFormat = ['json', 'geojson'].indexOf(req.query.format) >= 0 ? req.query.format : 'json';
-  if (req.outputFormat === 'geojson') {
-    req.fields.delete('contour');
-    req.fields.delete('centre');
-  }
-  next();
-}
-
-function formatCommune(req, commune) {
-  if (req.outputFormat === 'geojson') {
-    const geom = ['contour', 'centre'].indexOf(req.query.geometry) >= 0 ? req.query.geometry : 'centre';
-    return {
-      type: 'Feature',
-      properties: pick(commune, Array.from(req.fields)),
-      geometry: commune[geom]
-    };
-  } else {
-    return pick(commune, Array.from(req.fields));
-  }
-}
 
 const app = express();
 app.use(cors());
